app colors updated
