+++ lib/ui/core/themes/app_themes.dart
@@ -1,15 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 
-class AppColors {
-  static const Color primaryBlack = Color(0xFF1A1A1A);
-  static const Color darkGrey = Color(0xFF2A2A2A);
-  static const Color mediumGrey = Color(0xFF404040);
-  static const Color lightGrey = Color(0xFF808080);
-  static const Color yellow = Color(0xFFFFD700);
-  static const Color red = Color(0xFFFF4444);
-  static const Color white = Colors.white;
-}
+import 'app_colors.dart';
+
+
 
 class AppTheme {
   static ThemeData get darkTheme {
